pCmp: add data-element-scope attribute
This attribute describes an element's scope and can be used to
disambiguate amongst element types when a selection is made using an
overloaded element selector. See below for the motivating use case.

This patch also bugfixes the Plot object's select and unselect
functions, which previously selected all flow points and flow groups
matching with the matching data-code-line-number value even when they
were unrelated to the cursor's position. Appropriately-scoped selection
of these elements will be reinstated as needed later.